refactor: move load custom actions to core
